+++ CHANGELOG.md
@@ -5,9 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.21.2] - Unpublished
+## [0.21.2] - 2023-05-14
 
 ### Added
+
+- `split_coin_and_return` on SuiTransaction. Like split_coin_equally but splits and doesn't transfer back to self, instead
+  returns list where each element and list can be sent/transferred/used in subsequent transactions.
+- `public_transfer_object` on SuiTransaction, exposes 0x2::transfer::public_transfer
 
 ### Fixed